requirements.txt
Flask==2.2.5
gunicorn==21.2.0
requests==2.31.0
requests_oauthlib==1.3.0
mysql-connector-python==8.2.0